--- a/src/_data/ghost.js
+++ b/src/_data/ghost.js
@@ -3,15 +3,7 @@ const { chunk, cloneDeep } = require('lodash');
 const { resolve } = require('path');
 
 const fetchFromGhost = require('../../utils/ghost/fetch-from-ghost');
-const errorLogger = require('../../utils/error-logger');
-
-const { sourceApiUrl } = require('../../utils/ghost/api');
-const { siteURL, postsPerPage } = require('../../config');
-
-// Strip Ghost domain from urls
-const stripDomain = url => {
-  return url.replace(sourceApiUrl, '');
-};
+const { postsPerPage } = require('../../config');
 
 const getUniqueList = (arr, key) => [
   ...new Map(arr.map(item => [item[key], item])).values()
@@ -22,12 +14,12 @@ const piscina = new Piscina({
 });
 
 module.exports = async () => {
-  // Chunk posts and pages and process them in batches with
-  // a pool of workers
+  // Chunk raw Ghost posts and pages and process them in batches
+  // with a pool of workers to create posts and pages global data
   const batchSize = 200;
   const allPosts = await fetchFromGhost('posts');
   const allPages = await fetchFromGhost('pages');
-  const processedPosts = await Promise.all(
+  const posts = await Promise.all(
     chunk(allPosts, batchSize).map((batch, i, arr) =>
       piscina.run({
         batch,
@@ -42,7 +34,7 @@ module.exports = async () => {
       return arr.flat();
     })
     .catch(err => console.error(err));
-  const processedPages = await Promise.all(
+  const pages = await Promise.all(
     chunk(allPages, batchSize).map((batch, i, arr) =>
       piscina.run({
         batch,
@@ -58,43 +50,7 @@ module.exports = async () => {
     })
     .catch(err => console.error(err));
 
-  const posts = processedPosts.map(post => {
-    post.path = stripDomain(post.url);
-    post.primary_author.path = stripDomain(post.primary_author.url);
-    post.tags.forEach(tag => {
-      // Log and fix tag pages that point to 404 due to a Ghost error
-      if (tag.url.endsWith('/404/') && tag.visibility === 'public') {
-        errorLogger({ type: 'tag', name: tag.name });
-        tag.url = `${siteURL}/tag/${tag.slug}/`;
-      }
-
-      tag.path = stripDomain(tag.url);
-    });
-
-    // Log and fix author pages that point to 404 due to a Ghost error
-    if (post.primary_author.url.endsWith('/404/')) {
-      errorLogger({ type: 'author', name: post.primary_author.name });
-      post.primary_author.url = `${siteURL}/author/${post.primary_author.slug}/`;
-    }
-
-    post.primary_author.path = stripDomain(post.primary_author.url);
-
-    // Convert publish date into a Date object
-    post.published_at = new Date(post.published_at);
-
-    return post;
-  });
-
-  const pages = processedPages.map(page => {
-    page.path = stripDomain(page.url);
-    page.primary_author.path = stripDomain(page.primary_author.url);
-
-    // Convert publish date into a Date object
-    page.published_at = new Date(page.published_at);
-
-    return page;
-  });
-
+  // Create authors global data for author pages
   const authors = [];
   const primaryAuthors = getUniqueList(
     posts.map(post => post.primary_author),
@@ -113,7 +69,7 @@ module.exports = async () => {
           feature_image: post.feature_image,
           published_at: post.published_at,
           primary_author: post.primary_author,
-          tags: [post.tags[0]],
+          tags: [post.tags[0]], // Only include the first / primary tag
           image_dimensions: { ...post.image_dimensions }
         };
       });
@@ -136,6 +92,7 @@ module.exports = async () => {
     });
   });
 
+  // Create tags global data for tags pages
   const tags = [];
   const visibleTags = posts.reduce((arr, post) => {
     return [...arr, ...post.tags.filter(tag => tag.visibility === 'public')];
@@ -154,7 +111,7 @@ module.exports = async () => {
           feature_image: post.feature_image,
           published_at: post.published_at,
           primary_author: post.primary_author,
-          tags: [post.tags[0]],
+          tags: [post.tags[0]], // Only include the first / primary tag
           image_dimensions: { ...post.image_dimensions }
         };
       });
@@ -179,6 +136,7 @@ module.exports = async () => {
     });
   });
 
+  // Create popularTags global data to show at the top of tag pages
   const popularTags = [...allTags]
     .sort(
       (a, b) =>
@@ -212,6 +170,8 @@ module.exports = async () => {
         return feedObj;
       });
 
+  // Create feeds global data for the main, tags, and authors
+  // RSS feeds
   const feeds = [
     // Create custom collection for main RSS feed
     getCollectionFeeds([