--- a/src/_includes/layouts/feed.njk
+++ b/src/_includes/layouts/feed.njk
@@ -30,7 +30,8 @@
                 <link>{{ site.url + post.path }}</link>
                 <guid isPermaLink="false">{{ post.comment_id }}</guid>
                 <category>
-                    <![CDATA[ {{ post.tags[0].name }} ]]>
+                    {% set primaryTag = post.tags[0] %}
+                    <![CDATA[ {{ primaryTag.name }} ]]>
                 </category>
                 <dc:creator>
                     <![CDATA[ {{ post.primary_author.name }} ]]>