--- a/utils/get-image-dimensions.js
+++ b/utils/get-image-dimensions.js
@@ -1,9 +1,17 @@
 const probe = require('probe-image-size');
 const errorLogger = require('./error-logger');
+// Cache image dimensions we know will be repeated, like author profile and cover images
+const imageDimensionMap = {};
 
-const getImageDimensions = async (url, title) => {
+const getImageDimensions = async (url, title, cache) => {
   try {
+    if (cache && imageDimensionMap[url]) return imageDimensionMap[url];
+
     const { width, height } = await probe(url);
+
+    if (cache) {
+      imageDimensionMap[url] = { width, height };
+    }
 
     return {
       width,