--- a/utils/ghost/fetch-from-ghost.js
+++ b/utils/ghost/fetch-from-ghost.js
@@ -29,7 +29,8 @@ const fetchFromGhost = async endpoint => {
       });
 
     lastPage = ghostRes.meta.pagination.pages;
-    console.log(`Fetched ${endpoint} page ${currPage} of ${lastPage}...`);
+    if (ghostRes.length > 0)
+      console.log(`Fetched ${endpoint} page ${currPage} of ${lastPage}...`);
     currPage = ghostRes.meta.pagination.next;
 
     ghostRes.forEach(obj => data.push(obj));