Checklist:

<!-- Please follow this checklist and put an x in each of the boxes, like this: [x]. It will ensure that our team takes your pull request seriously. -->

- [x] I have read [freeCodeCamp's contribution guidelines](https://contribute.freecodecamp.org).
- [x] My pull request has a [descriptive title](https://contribute.freecodecamp.org/#/how-to-open-a-pull-request?id=prepare-a-good-pr-title) (**not** a vague title like `Update index.md`)

<!--If your pull request closes a GitHub issue, replace the XXXXX below with the issue number.-->

Closes #XXXXX

<!-- Feel free to add any additional description of changes below this line -->
This PR does a few things, mainly optimizing memory usage and refactoring the build process into distinct fetch-from-Ghost and process-Ghost-data stages.

To optimize memory usage, as many unused properties are removed from the Ghost post / page objects as possible, and caching for image dimensions are limited to author profile images -- which have a high probability of being used multiple times throughout the site -- rather than all images outside of the ones in the Ghost HTML.

Currently, as pages of 200 posts are fetched from the Ghost API, they're immediately processed (getting image dimensions, generating data for AMP  pages, etc.) using several `await Promise.all( arr.map... )` loops. This ties up the main Node thread, and the next page of 200 posts aren't fetched until the current batch is finished processing. The long processing time between fetching pages of data from Ghost has also led to some failed builds due to timeouts.

To improve build times, the build process has been split up into two separate steps. The first is to get all the posts / pages from Ghost as quickly as possible. The next step is to use [Piscina](https://www.npmjs.com/package/piscina) to manage Node worker pools and offload batches of post / page processing to unused CPU threads. Once all the batches of posts / pages are complete, 11ty can do its thing.

The downside is that all builds will require more RAM. Also, building smaller News sites (< 1,000 articles) may take around the same amount of time, if not slightly longer.

But the upside is that building English News (8,673 articles) should be much faster.

Here are some average build times on a DO droplet located in Singapore with 8 vCPUs and 16GB of RAM:

|  Language  | Published articles | Files written |                 Current build time                 |              Multithreaded build time             |
|:----------:|:------------------:|:-------------:|:--------------------------------------------------:|:-------------------------------------------------:|
|   English  |        8,673       |     23,491    | 1844.99 seconds / 30.75 minutes (78.5ms each file) | 876.92 seconds / 14.62 minutes (37.3ms each file) |
|   Chinese  |         848        |     2,420     |  270.06 seconds / 4.5 minutes (111.6ms each file)  |  139.16 seconds / 2.32 minutes (57.5ms each file) |
| Portuguese |         253        |      744      |   56.62 seconds / 0.94 minutes (76.1ms each file)  |  49.90 seconds / 0.83 minutes (67.1ms each file)  |

It's a small sample size, and I only did one test build each for the data above, but I've seen a similar trend on my machine (2 cores / 4 threads, 16GB of RAM) while working on this.