@raisedadead 2022-05-11
Hey @scissorsneedfoodtoo

You did a fantastic job here. With this change, the relatively underpowered CI virtual machines are giving us better build times. We have shaved a good 6-8 mins off a build on English news.

Kudos! 💯🚀🎉