@scissorsneedfoodtoo 2022-05-12
Thanks @raisedadead! I'm glad to hear that build times have improved. Hopefully we'll see fewer failed builds as well.

I've been doing some additional testing with the latest beta version of 11ty where they also made some optimizations on their end. Once those changes are landed in a stable release, we should see even better build times.

I'll also continue looking into the serverless plugin, which now seems to work with AWS Lambda functions. That may be a path forward sometime in the future.